# docs/automation/openapi-v1.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "paths": {
...
    "/api/board/{section}": {
      "post": {
        "operationId": "PostApiBoard",
        "parameters": [
          {
            "name": "section",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BoardSection"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemTitleRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/board/{section}/{itemId}": {
      "put": {
        "operationId": "PutApiBoard",
        "parameters": [
          {
            "name": "section",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BoardSection"
            },
            "x-position": 1
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemTitleRequest"
              }
            }
          },
          "required": true,
          "x-position": 3
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "DeleteApiBoard",
        "parameters": [
          {
            "name": "section",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BoardSection"
            },
            "x-position": 1
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/board/{section}/{itemId}/archive": {
      "post": {
        "operationId": "PostApiBoardArchive",
        "parameters": [
          {
            "name": "section",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BoardSection"
            },
            "x-position": 1
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/board/{section}/{itemId}/unarchive": {
      "post": {
        "operationId": "PostApiBoardUnarchive",
        "parameters": [
          {
            "name": "section",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BoardSection"
            },
            "x-position": 1
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/board/{section}/{itemId}/toggle": {
      "post": {
        "operationId": "PostApiBoardToggle",
        "parameters": [
          {
            "name": "section",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BoardSection"
            },
            "x-position": 1
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/board/habits/{itemId}/increment": {
      "post": {
        "operationId": "PostApiBoardHabitsIncrement",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/board/habits/{itemId}/decrement": {
      "post": {
        "operationId": "PostApiBoardHabitsDecrement",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/board/habits/{itemId}": {
      "put": {
        "operationId": "PutApiBoardHabits",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HabitUpdateRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/board/todos/{itemId}": {
      "put": {
        "operationId": "PutApiBoardTodos",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoUpdateRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/board/dailies/{itemId}": {
      "put": {
        "operationId": "PutApiBoardDailies",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyUpdateRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/board/dailies/{itemId}/complete-for-date": {
      "post": {
        "operationId": "PostApiBoardDailiesCompleteForDate",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyCompleteForDateRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/board/dailies/{itemId}/skip-for-date": {
      "post": {
        "operationId": "PostApiBoardDailiesSkipForDate",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailySkipForDateRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BoardSection": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Habit",
          "Daily",
          "Todo"
        ],
        "enum": [
          0,
          1,
          2
        ]
      },
...
      "HabitResetPeriod": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Daily",
          "Weekly",
          "Monthly"
        ],
        "enum": [
          0,
          1,
          2
        ]
      },
      "DailyRepeatType": {
        "type": "integer",
        "description": "How often a daily, or its checklist due logic, is scheduled to repeat.",
        "x-enumNames": [
          "Daily",
          "Weekly",
          "Monthly",
          "Yearly"
        ],
        "enum": [
          0,
          1,
          2,
          3
        ]
      },
...
      "ItemTitleRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string"
          },
          "itemId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          }
        }
      },
      "HabitUpdateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "string",
            "nullable": true
          },
          "trackPlus": {
            "type": "boolean"
          },
          "trackMinus": {
            "type": "boolean"
          },
          "resetPeriod": {
            "$ref": "#/components/schemas/HabitResetPeriod"
          },
          "counter": {
            "type": "integer",
            "format": "int32"
          },
          "negativeCounter": {
            "type": "integer",
            "format": "int32"
          },
          "checklistJson": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        }
      },
      "TodoUpdateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "string",
            "nullable": true
          },
          "checklistJson": {
            "type": "string",
            "nullable": true
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "sortOrder": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        }
      },
      "DailyUpdateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "repeat": {
            "$ref": "#/components/schemas/DailyRepeatType"
          },
          "repeatInterval": {
            "type": "integer",
            "format": "int32"
          },
          "checklistJson": {
            "type": "string",
            "nullable": true
          },
          "counter": {
            "type": "integer",
            "format": "int32"
          },
          "sortOrder": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "weekdays": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "DailyCompleteForDateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "completedOn": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "DailySkipForDateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "skippedOn": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
    }
  }
}